fixed a bug that was triggered on 64 bit machines
git-svn-id: file:///home/git/hedgehog.fhcrc.org/bioconductor/trunk/madman/Rpacks/genefilter@26013 bc3139a8-67e5-0310-9ffc-ced21a209358

+++ src/nd.c
@@ -26,7 +26,7 @@
 #if defined(USING_R) /*( R-specific stuff */
 
 #define  S_CDECL
-typedef long RSInt;
+typedef int RSInt;
 
 #ifdef HAVE_CONFIG_H
 # include <config.h>
